Added map importing from the map folder. Separated layers to new file layers.py Screens and menus are passed the state engine, rather than display and session. Added enter and exit functions to screens. Added player manager class. Removed army as a value belonging to player.
Map is now built in two steps. WorldSetup creates a MapSetup which makes a 2d array of EntitySeeds.
Entity seeds can be passed to WorldDict.entity_from_seed() to instanciate the entity.
This way the map only gets instantiated when it is actually loaded, but the map data is visible in world_setup.